fixed the schema, it builds from  csv.

+++ data/schema.sql
@@ -7,7 +7,7 @@ CREATE TABLE camNames(
     id INT AUTO_INCREMENT NOT NULL,
     year YEAR NOT NULL,
     name VARCHAR(30) NOT NULL,
-    percentage INT,
+    percentage FLOAT(8,7),
     sex VARCHAR(30),
     PRIMARY KEY(id)
 );
